refactor #4 Moved the get_markov_blanket method
Fixes #4
